+++ lib/nextjs-themes/src/client/color-switch/color-switch.test.tsx
@@ -1,6 +1,6 @@
 import { act, cleanup, fireEvent, render, renderHook, screen } from "@testing-library/react";
 import { ColorSwitch } from "./color-switch";
-import { useTheme } from "../../store";
+import { useTheme } from "../../hooks";
 
 describe("color-switch", () => {
   afterEach(cleanup);