+++ lib/nextjs-themes/src/client/force-color-scheme/force-color-scheme.tsx
@@ -1,16 +1,16 @@
 "use client";
 import * as React from "react";
 import { useEffect } from "react";
-import type { ColorSchemeType } from "../../store";
-import { useTheme } from "../../store";
+import type { ColorSchemeType } from "../../constants";
+import { useTheme } from "../../hooks";
 
 export function ForceColorScheme(props: { colorScheme: ColorSchemeType }) {
-  const [setForcedColorScheme] = useTheme(state => [state.setForcedColorScheme]);
+  const { setForcedColorScheme } = useTheme();
   useEffect(() => {
     setForcedColorScheme(props.colorScheme);
     return () => {
       setForcedColorScheme(undefined);
     };
-  }, [props.colorScheme, setForcedColorScheme]);
+  }, [props.colorScheme]);
   return null;
 }