+++ lib/nextjs-themes/src/client/theme-switcher/theme-switcher.tsx
@@ -1,20 +1,49 @@
 import * as React from "react";
 import { useEffect } from "react";
-import type { ColorSchemeType } from "../../store";
-import { useTheme } from "../../store";
-import { resolveTheme } from "../../utils";
-import { StorageType } from "persist-and-sync";
+import { resolveTheme, parseState, encodeState } from "../../utils";
+import { ColorSchemeType, DEFAULT_ID, ThemeStoreType, initialState } from "../../constants";
+import useRGS, { SetStateAction } from "r18gs";
 
+/** todo - set persistance and cookies */
 export interface ThemeSwitcherProps {
   forcedTheme?: string;
   forcedColorScheme?: ColorSchemeType;
   targetSelector?: string;
   themeTransition?: string;
-  /**
-   * defaultValue `"cookies"`
-   * set storage to `localStorage` or `sessionsStorage` when using only client side or when you must avoid using cookies
-   */
-  storage?: StorageType;
+}
+
+function useMediaQuery(setThemeState: SetStateAction<ThemeStoreType>) {
+  React.useEffect(() => {
+    // set event listener for media
+    const media = matchMedia("(prefers-color-scheme: dark)");
+    const updateSystemColorScheme = () => {
+      setThemeState(state => ({ ...state, systemColorScheme: media.matches ? "dark" : "light" }));
+    };
+    updateSystemColorScheme();
+    media.addEventListener("change", updateSystemColorScheme);
+    return () => {
+      media.removeEventListener("change", updateSystemColorScheme);
+    };
+  }, [setThemeState]);
+}
+
+let tInit = 0;
+function useLoadSyncedState(setThemeState: SetStateAction<ThemeStoreType>, targetSelector?: string) {
+  React.useEffect(() => {
+    tInit = Date.now();
+    const key = targetSelector ?? DEFAULT_ID;
+    const storedState = parseState(localStorage.getItem(key));
+    // @ts-expect-error -- using only as a partial
+    delete storedState.systemColorScheme;
+    setThemeState(state => ({ ...state, ...storedState }));
+    const storageListener = (e: StorageEvent) => {
+      if (e.key === key) setThemeState(state => ({ ...state, ...parseState(e.newValue) }));
+    };
+    window.addEventListener("storage", storageListener);
+    return () => {
+      window.removeEventListener("storage", storageListener);
+    };
+  }, [targetSelector]);
 }
 
 export interface DataProps {
@@ -34,10 +63,11 @@ export interface UpdateProps {
 
 const updateDOM = (
   { resolvedTheme, resolvedColorScheme, resolvedColorSchemePref, th }: UpdateProps,
-  isSystemDark: boolean,
   targetSelector?: string,
 ) => {
-  [document.querySelector(targetSelector || "#nextjs-themes"), document.documentElement].forEach(target => {
+  const target = document.querySelector(targetSelector || `#${DEFAULT_ID}`);
+  /** don't apply theme to documentElement for localized targets */
+  [target, targetSelector && target ? null : document.documentElement].forEach(target => {
     /** ensuring that class 'dark' is always present when dark color scheme is applied to support Tailwind  */
     if (target)
       target.className = `${resolvedColorScheme} theme-${resolvedTheme} th-${th} csp-${resolvedColorSchemePref}`;
@@ -46,14 +76,14 @@ const updateDOM = (
     target?.setAttribute("data-color-scheme", resolvedColorScheme);
     target?.setAttribute("data-csp", resolvedColorSchemePref); /** color-scheme-preference */
   });
-
-  /** store system preference for computing data-theme on server side */
-  document.cookie = `data-color-scheme-system=${isSystemDark ? "dark" : "light"}`;
+  const shouldCreateCookie = target?.getAttribute("data-nth") === "next";
+  return shouldCreateCookie;
 };
 
 const disableAnimation = (themeTransition = "none") => {
   const css = document.createElement("style");
-  const transition = `transition: ${themeTransition} !important;`;
+  /** split by ';' to prevent CSS injection */
+  const transition = `transition: ${themeTransition.split(";")[0]} !important;`;
   css.appendChild(
     document.createTextNode(`*{-webkit-${transition}-moz-${transition}-o-${transition}-ms-${transition}${transition}}`),
   );
@@ -74,39 +104,23 @@ const disableAnimation = (themeTransition = "none") => {
  * Please note that you need to add "use client" on top of the component in which you are using this hook.
  */
 export function useThemeSwitcher(props: ThemeSwitcherProps) {
-  const [setStorage, ...depArray] = useTheme(state => [
-    state.setStorage,
-    state.theme,
-    state.darkTheme,
-    state.lightTheme,
-    state.colorSchemePref,
-    state.forcedColorScheme,
-    state.forcedTheme,
-  ]);
+  const [themeState, setThemeState] = useRGS<ThemeStoreType>(props.targetSelector ?? DEFAULT_ID, initialState);
 
+  useMediaQuery(setThemeState);
+  useLoadSyncedState(setThemeState, props.targetSelector);
   useEffect(() => {
-    setStorage(props.storage ?? "cookies");
-  }, [props.storage]);
+    const restoreTransitions = disableAnimation(props.themeTransition);
 
-  useEffect(() => {
-    const themeState = useTheme.getState();
-    const media = matchMedia("(prefers-color-scheme: dark)");
-    const updateTheme = () => {
-      const restoreTransitions = disableAnimation(props.themeTransition);
-
-      const resolvedData = resolveTheme(media.matches, themeState, props);
-      themeState.setResolved(resolvedData);
-      updateDOM(resolvedData, media.matches, props.targetSelector);
-
-      restoreTransitions();
-    };
-
-    media.addEventListener("change", updateTheme);
-    updateTheme();
-    return () => {
-      media.removeEventListener("change", updateTheme);
-    };
-  }, [props.forcedColorScheme, props.forcedTheme, props.targetSelector, ...depArray]);
+    const resolvedData = resolveTheme(themeState, props);
+    const shouldCreateCookie = updateDOM(resolvedData, props.targetSelector);
+    if (tInit < Date.now() - 300) {
+      const stateStr = encodeState(themeState);
+      const key = props.targetSelector || DEFAULT_ID;
+      localStorage.setItem(key, stateStr);
+      if (shouldCreateCookie) document.cookie = `${key}=${stateStr}; max-age=31536000; SameSite=Strict;`;
+    }
+    restoreTransitions();
+  }, [props, themeState]);
 }
 
 /**