+++ lib/nextjs-themes/src/index.ts
@@ -1,4 +1,5 @@
 "use client";
 // client component exports
 export * from "./client";
-export * from "./store";
+export * from "./hooks";
+export * from "./constants";