+++ lib/nextjs-themes/src/utils.ts
@@ -1,12 +1,14 @@
 import { ThemeSwitcherProps, UpdateProps } from "./client";
-import type { ThemeStoreType } from "./store";
+import { ColorSchemeType, ThemeStoreType, initialState } from "./constants";
 
-export function resolveTheme(isSystemDark: boolean, state?: ThemeStoreType, props?: ThemeSwitcherProps): UpdateProps {
+export function resolveTheme(state?: ThemeStoreType, props?: ThemeSwitcherProps): UpdateProps {
   const resolvedForcedTheme = props?.forcedTheme === undefined ? state?.forcedTheme : props.forcedTheme;
   const resolvedForcedColorScheme =
     props?.forcedColorScheme === undefined ? state?.forcedColorScheme : props.forcedColorScheme;
   const resolvedColorSchemePref =
     (resolvedForcedColorScheme === undefined ? state?.colorSchemePref : resolvedForcedColorScheme) || "";
+
+  const isSystemDark = state?.systemColorScheme === "dark";
 
   let resolvedColorScheme: "dark" | "light" = isSystemDark ? "dark" : "light";
   let resolvedTheme = resolvedForcedTheme === undefined ? state?.theme || "" : resolvedForcedTheme;
@@ -34,4 +36,16 @@ export function getResolvedTheme() {
 
 export function getResolvedColorScheme() {
   return document.documentElement.getAttribute("data-color-scheme");
+}
+
+export function encodeState(themeState: ThemeStoreType) {
+  const { colorSchemePref, systemColorScheme, darkTheme, lightTheme, theme } = themeState;
+  return [colorSchemePref, systemColorScheme, darkTheme, lightTheme, theme].join(",");
+}
+
+export function parseState(str?: string | null): ThemeStoreType {
+  if(!str) return initialState;
+  type StrSplitType = [ColorSchemeType, "dark" | "light", string, string, string];
+  const [colorSchemePref, systemColorScheme, darkTheme, lightTheme, theme] = str.split(",") as StrSplitType;
+  return { colorSchemePref, systemColorScheme, darkTheme, lightTheme, theme };
 }