+++ examples/tailwind/CHANGELOG.md
@@ -1,5 +1,13 @@
 # tailwind
 
+## 0.1.8
+
+### Patch Changes
+
+- Updated dependencies [9f30638]
+- Updated dependencies [9f30638]
+  - nextjs-themes@3.0.0
+
 ## 0.1.7
 
 ### Patch Changes